Move memory regions out into a separate header for clarity
